Minor fixes and cleanup; Rename ClassParser to more explainable AnnotationParser; Improved test coverage
