+++ src/Helper/Reflection/AnnotationParser.php
@@ -8,7 +8,7 @@ namespace Faulancer\Helper\Reflection;
  * @package Faulancer\Reflection
  * @author  Florian Knapp <office@florianknapp.de>
  */
-class ClassParser extends \ReflectionClass
+class AnnotationParser extends \ReflectionClass
 {
 
     /** @var string */
@@ -37,7 +37,6 @@ class ClassParser extends \ReflectionClass
         }
 
         return $result;
-
     }
 
     /**
@@ -49,22 +48,27 @@ class ClassParser extends \ReflectionClass
     private function extractValues($name, $method)
     {
         $arr       = [];
+        $vars      = [];
         $methodDoc = new \ReflectionMethod($this->className, $method);
 
         preg_match('|@' . $name . '(.*)|', $methodDoc->getDocComment(), $arr);
 
-        $var = [];
+        if (empty($arr)) {
+            return [];
+        }
 
-        preg_match('|\((.*)\)|', $arr[1], $var);
+        preg_match('|\((.*)\)|', $arr[1], $vars);
 
         $param = str_replace(
             ['"', ',', ' ', '+'],
             ['', '&', '', '___'],
-            $var[1]
+            $vars[1]
         );
 
+        // Parse uri conform string into key/value pairs
         parse_str($param, $result);
 
+        // Add class and action to result variable
         $result['action'] = $method;
         $result['class']  = $this->className;
