--- tests/Unit/TestTranslation.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Faulancer\Test\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-/**
- * Class TestTranslation
- * @package Faulancer\Test\Unit
- */
-class TestTranslation extends TestCase
-{
-
-    public function testTranslation()
-    {
-        
-    }
-
-}
